pyproject.toml
[project]
name = "forum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.0.1",
    "django==5.2.*",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "markdown>=3.9",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]
